Okay, finally the indexer is now on the DB.
Updates are that we now have the forward and inverted index on the database, along with the offline version of the inverted index.
Besides, I have added the DBSearch function so we can search on the database, the offline search is still there but I don't think we'll need it.